+++ meta/packages.ts
@@ -1,4 +1,8 @@
 export const packages = [
   { name: 'mixte' },
   { name: 'use' },
+  {
+    name: 'use/vite',
+    external: ['unplugin-auto-import/types'],
+  },
 ];